move global methods to own file
